Move config watcher wiring out of DaemonServer
The ConfigWatcher singleton was wired entirely through DaemonServer: its
start()/stop() lifecycle, the six reload-reaction callbacks, plus dead
lastConfigFingerprint/refreshConfigFromSources accessors. Move the wiring to the
daemon startup/shutdown sequences (getConfigWatcher().start/stop), relocate the
shared conversation-eviction logic to conversation-store (evictConversationsForReload,
also used on provider-credential change), and define the IDENTITY.md broadcast
reaction in lifecycle. The trivial sounds/avatar/config reactions are passed
straight through as the publish* functions.

ConfigWatcher's API is unchanged. The dead lastConfigFingerprint/
refreshConfigFromSources accessors (no callers) and refreshConversationsForProviderChange
are removed.
