+++ assistant/src/daemon/lifecycle.ts
@@ -1,3 +1,4 @@
+import { existsSync, readFileSync } from "node:fs";
 import { join } from "node:path";
 
 import { config as dotenvConfig } from "dotenv";
@@ -10,7 +11,11 @@ import { TwilioConversationRelayProvider } from "../calls/twilio-provider.js";
 import { setVoiceBridgeDeps } from "../calls/voice-session-bridge.js";
 import { initFeatureFlagOverrides } from "../config/assistant-feature-flags.js";
 import { setIngressPublicBaseUrl, validateEnv } from "../config/env.js";
-import { loadConfig, mergeDefaultWorkspaceConfig } from "../config/loader.js";
+import {
+  getConfig,
+  loadConfig,
+  mergeDefaultWorkspaceConfig,
+} from "../config/loader.js";
 import { seedInferenceProfiles } from "../config/seed-inference-profiles.js";
 import { reconcileFlagGatedProfiles } from "../config/sync-gated-profiles.js";
 import { startCes } from "../credential-execution/ces-runtime.js";
@@ -55,6 +60,7 @@ import {
   startConsentRefresh,
   stopConsentRefresh,
 } from "../platform/consent-cache.js";
+import { syncIdentityNameToPlatform } from "../platform/sync-identity.js";
 import { ensurePromptFiles } from "../prompts/system-prompt.js";
 import { runProviderConnectionsBackfill } from "../providers/inference/backfill.js";
 import { broadcastMessage } from "../runtime/assistant-event-hub.js";
@@ -67,8 +73,11 @@ import { warmLocalGuardianPrincipalCache } from "../runtime/local-actor-identity
 import { recoverInterruptedImport } from "../runtime/migrations/vbundle-streaming-importer.js";
 import { registerSecretsDeps } from "../runtime/routes/secrets-deps.js";
 import {
+  publishAvatarChanged,
   publishConfigChanged,
   publishConversationListChanged,
+  publishIdentityChanged,
+  publishSoundsConfigUpdated,
 } from "../runtime/sync/resource-sync-events.js";
 import { recoverStaleSchedules } from "../schedule/schedule-recovery.js";
 import { startScheduler } from "../schedule/scheduler.js";
@@ -81,6 +90,7 @@ import {
   ensureDataDir,
   getDotEnvPath,
   getWorkspaceDir,
+  getWorkspacePromptPath,
 } from "../util/platform.js";
 import { APP_VERSION } from "../version.js";
 import {
@@ -93,6 +103,8 @@ import { startWorkspaceHeartbeatService } from "../workspace/heartbeat-service.j
 import { WORKSPACE_MIGRATIONS } from "../workspace/migrations/registry.js";
 import { runWorkspaceMigrations } from "../workspace/migrations/runner.js";
 import { startAppSourceWatcher } from "./app-source-watcher.js";
+import { getConfigWatcher } from "./config-watcher.js";
+import { evictConversationsForReload } from "./conversation-store.js";
 import { writePid } from "./daemon-control.js";
 import {
   evaluateDiskPressureNow,
@@ -102,6 +114,7 @@ import {
 import { startEventLoopWatchdog } from "./event-loop-watchdog.js";
 import { initializePlugins } from "./external-plugins-bootstrap.js";
 import { backfillSlackInjectionTemplates } from "./handlers/config-slack-channel.js";
+import { parseIdentityFields } from "./handlers/identity.js";
 import { installAssistantSymlink } from "./install-symlink.js";
 import {
   maybeRebuildMemoryV2Concepts,
@@ -171,6 +184,27 @@ export function stopDiskPressureGuardForLifecycle(): void {
     diskPressureStartupSampleTimer = null;
   }
   stopDiskPressureGuard();
+}
+
+/**
+ * Re-read IDENTITY.md and broadcast the parsed fields to clients, best-effort
+ * syncing the assistant name to the platform record. Wired as the config
+ * watcher's IDENTITY.md-change reaction.
+ */
+function broadcastIdentityChange(): void {
+  try {
+    const identityPath = getWorkspacePromptPath("IDENTITY.md");
+    const content = existsSync(identityPath)
+      ? readFileSync(identityPath, "utf-8")
+      : "";
+    const fields = parseIdentityFields(content);
+    publishIdentityChanged(fields);
+    if (fields.name) {
+      syncIdentityNameToPlatform(fields.name);
+    }
+  } catch (err) {
+    log.error({ err }, "Failed to broadcast identity change");
+  }
 }
 
 // Entry point for the daemon process itself
@@ -591,6 +625,20 @@ export async function runDaemon(): Promise<void> {
   await server.start();
   log.info("Daemon startup: DaemonServer started");
 
+  // Watch workspace files (config, prompts, skills, sounds, avatar) for changes
+  // and react: evict conversations so the next turn rebuilds against the new
+  // config, and broadcast the relevant resource-changed events to clients.
+  const configWatcher = getConfigWatcher();
+  configWatcher.initFingerprint(getConfig());
+  configWatcher.start(
+    evictConversationsForReload,
+    broadcastIdentityChange,
+    publishSoundsConfigUpdated,
+    publishAvatarChanged,
+    publishConfigChanged,
+    () => refreshSkillCapabilityMemories(getConfig()),
+  );
+
   // Watch app source directories so edits recompile + refresh surfaces across
   // all conversations.
   startAppSourceWatcher();
@@ -900,8 +948,7 @@ export async function runDaemon(): Promise<void> {
   // the running routes. They're module-level state, so they're effective
   // even when the HTTP server failed to bind (IPC clients still work).
   registerSecretsDeps({
-    onProviderCredentialsChanged: () =>
-      server.refreshConversationsForProviderChange(),
+    onProviderCredentialsChanged: () => evictConversationsForReload(),
   });
 
   // Fire-and-forget: Qdrant init and memory worker startup run concurrently
