+++ assistant/src/daemon/server.ts
@@ -7,30 +7,18 @@ import { syncIdentityNameToPlatform } from "../platform/sync-identity.js";
 import { initializeProviders } from "../providers/registry.js";
 import { broadcastMessage } from "../runtime/assistant-event-hub.js";
 import { getSigningKeyFingerprint } from "../runtime/auth/token-service.js";
-import {
-  publishAvatarChanged,
-  publishConfigChanged,
-  publishIdentityChanged,
-  publishSoundsConfigUpdated,
-} from "../runtime/sync/resource-sync-events.js";
 import { getSubagentManager } from "../subagent/index.js";
 import { getLogger } from "../util/logger.js";
 import { getWorkspacePromptPath } from "../util/platform.js";
-import { getConfigWatcher } from "./config-watcher.js";
 import { Conversation } from "./conversation.js";
 import { ConversationEvictor } from "./conversation-evictor.js";
-import {
-  conversationEntries,
-  deleteConversation,
-  getConversationMap,
-} from "./conversation-registry.js";
+import { getConversationMap } from "./conversation-registry.js";
 import {
   getOrCreateConversation as getOrCreateActiveConversation,
   initConversationLifecycle,
 } from "./conversation-store.js";
 import { parseIdentityFields } from "./handlers/identity.js";
 import type { ConversationCreateOptions } from "./handlers/shared.js";
-import { refreshSkillCapabilityMemories } from "./skill-memory-refresh.js";
 
 const log = getLogger("server");
 
@@ -52,9 +40,6 @@ export class DaemonServer {
   private sharedRequestTimestamps: number[] = [];
   private evictor: ConversationEvictor;
 
-  // Composed subsystems
-  private configWatcher = getConfigWatcher();
-
   constructor() {
     this.evictor = new ConversationEvictor(getConversationMap());
     getSubagentManager().sharedRequestTimestamps = this.sharedRequestTimestamps;
@@ -75,24 +60,6 @@ export class DaemonServer {
     };
   }
 
-  private broadcastIdentityChanged(): void {
-    try {
-      const identityPath = getWorkspacePromptPath("IDENTITY.md");
-      const content = existsSync(identityPath)
-        ? readFileSync(identityPath, "utf-8")
-        : "";
-      const fields = parseIdentityFields(content);
-      publishIdentityChanged(fields);
-
-      // Best-effort sync of the assistant name to the platform record.
-      if (fields.name) {
-        syncIdentityNameToPlatform(fields.name);
-      }
-    } catch (err) {
-      log.error({ err }, "Failed to broadcast identity change");
-    }
-  }
-
   /** Best-effort sync of the IDENTITY.md name to the platform record. */
   private syncIdentityToPlatform(): void {
     try {
@@ -109,35 +76,13 @@ export class DaemonServer {
     }
   }
 
-  private broadcastConfigChanged(): void {
-    publishConfigChanged();
-  }
-
-  private broadcastSoundsConfigUpdated(): void {
-    publishSoundsConfigUpdated();
-  }
-
-  private broadcastAvatarUpdated(): void {
-    publishAvatarChanged();
-  }
-
   // ── Server lifecycle ────────────────────────────────────────────────
 
   async start(): Promise<void> {
     const config = getConfig();
     await initializeProviders(config);
-    this.configWatcher.initFingerprint(config);
 
     this.evictor.start();
-
-    this.configWatcher.start(
-      () => this.evictConversationsForReload(),
-      () => this.broadcastIdentityChanged(),
-      () => this.broadcastSoundsConfigUpdated(),
-      () => this.broadcastAvatarUpdated(),
-      () => this.broadcastConfigChanged(),
-      () => refreshSkillCapabilityMemories(getConfig()),
-    );
 
     this.syncIdentityToPlatform();
 
@@ -148,7 +93,6 @@ export class DaemonServer {
     getSubagentManager().disposeAll();
     disposeAcpSessionManager();
     this.evictor.stop();
-    this.configWatcher.stop();
 
     log.info("Daemon server stopped");
   }
@@ -161,42 +105,6 @@ export class DaemonServer {
       version: daemonVersion,
       keyFingerprint: getSigningKeyFingerprint(),
     });
-  }
-
-  private evictConversationsForReload(): void {
-    const subagentManager = getSubagentManager();
-    for (const [id, conversation] of conversationEntries()) {
-      if (!conversation.isProcessing()) {
-        subagentManager.abortAllForParent(id);
-        conversation.dispose();
-        deleteConversation(id);
-        this.evictor.remove(id);
-      } else {
-        conversation.markStale();
-      }
-    }
-  }
-
-  get lastConfigFingerprint(): string {
-    return this.configWatcher.lastFingerprint;
-  }
-
-  set lastConfigFingerprint(value: string) {
-    this.configWatcher.lastFingerprint = value;
-  }
-
-  async refreshConfigFromSources(): Promise<boolean> {
-    const changed = await this.configWatcher.refreshConfigFromSources();
-    if (changed) this.evictConversationsForReload();
-    return changed;
-  }
-
-  /**
-   * Provider instances are captured when conversations are created, so a key
-   * change must evict or mark them stale before the next turn.
-   */
-  refreshConversationsForProviderChange(): void {
-    this.evictConversationsForReload();
   }
 
   /**
